requirements.txt
Django==4.2.4
djangorestframework==3.14.0

isort==5.12.0
typing-extensions==4.7.1
